refactor uofg-header to split up source into separate components as well as easier variant creation
